WebFluxTest demonstration
Let's start writing our test for our simple demo. We want to get a sushi
place whose price per person is lower than 40$. Our test doesn't pass as
we haven't implemented the feature yet but we can show how easy it is
to test a WebFlux controller with the new `WebTestClient`.