Flux handling demonstration
Let's assume a RestaurantLookupService that returns all the restaurants
in our area. The filter operator on Reactor allows us to implement the
requested feature but unfortunately our test does not start anymore as
it uses slicing.

We can mock the lookup service and returns sample data. But we still
haven't a way to get those restaurants and getting them all is too
limited.