MongoDB demonstration
Rather than retrieving all restaurants and doing the filtering ourselves
we can obviously use a store for that. In this commit, we add the
reactive MongoDB support which allows us to get this 'findAll()' method
out-of-the-box.

But we can also create our own search query and use it in our controller
instead.