Remote service demonstration
Let's assume that we have a reservation service running on port 8081.

We want to get the first 3 restaurants that are available. Reactor helps
us to combine data with a HTTP remote call to check availability, and
the take operation limits the requests transparently for us.

The remote service is slow and unreliable but it doesn't matter and we
can stream the response as it comes, try:

http :8080/restaurants/available --stream